+++ pipes-and-filters/src/main/scala/asso/benchmarking/Benchmarker.scala
@@ -1,4 +1,4 @@
-package asso
+package asso.benchmarking
 
 object Benchmarker {
